pyproject.toml
[project]
name = "ollama-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.5.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
